+++ src/components/Accounts/Actions/ActionModal.js
@@ -2,13 +2,10 @@ import _ from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { FormattedMessage } from 'react-intl';
-import {
-  Field,
-  reduxForm,
-  change,
-  formValueSelector,
-} from 'redux-form';
-import { connect } from 'react-redux';
+import { Field } from 'react-final-form';
+import setFieldData from 'final-form-set-field-data';
+
+import stripesFinalForm from '@folio/stripes/final-form';
 import SafeHTMLMessage from '@folio/react-intl-safe-html';
 import {
   Row,
@@ -16,53 +13,28 @@ import {
   Button,
   TextArea,
   Modal,
+  NoValue,
   TextField,
   Checkbox,
   Select,
 } from '@folio/stripes/components';
 
 import { calculateSelectedAmount } from '../accountFunctions';
+
 import css from './PayWaive.css';
-
-const validate = (values, props) => {
-  const {
-    accounts,
-    action,
-    commentRequired
-  } = props;
-
-  const selected = calculateSelectedAmount(accounts);
-  const errors = {};
-
-  if (!parseFloat(values.amount)) {
-    errors.amount = <FormattedMessage id="ui-users.accounts.error.field" />;
-  }
-  if (parseFloat(values.amount) <= 0) {
-    errors.amount = <FormattedMessage id={`ui-users.accounts.${action}.error.amount`} />;
-  }
-  if (!values.method) {
-    errors.method = <FormattedMessage id={`ui-users.accounts.${action}.error.select`} />;
-  }
-  if (commentRequired && !values.comment) {
-    errors.comment = <FormattedMessage id="ui-users.accounts.error.comment" />;
-  }
-  if (parseFloat(values.amount) > parseFloat(selected)) {
-    errors.amount = <FormattedMessage id={`ui-users.accounts.${action}.error.exceeds`} />;
-  }
-
-  return errors;
-};
 
 class ActionModal extends React.Component {
   static propTypes = {
+    form: PropTypes.object.isRequired,
     onClose: PropTypes.func,
     handleSubmit: PropTypes.func,
     open: PropTypes.bool,
     accounts: PropTypes.arrayOf(PropTypes.object),
     data: PropTypes.arrayOf(PropTypes.object),
     balance: PropTypes.string,
+    totalPaidAmount: PropTypes.string,
+    owedAmount: PropTypes.string,
     submitting: PropTypes.bool,
-    invalid: PropTypes.bool,
     pristine: PropTypes.bool,
     reset: PropTypes.func,
     commentRequired: PropTypes.bool,
@@ -70,13 +42,40 @@ class ActionModal extends React.Component {
     feefines: PropTypes.arrayOf(PropTypes.object),
     label: PropTypes.string,
     action: PropTypes.string,
-    intl: PropTypes.object,
-    currentValues: PropTypes.object,
-    dispatch: PropTypes.func,
+    intl: PropTypes.object.isRequired,
+    checkAmount: PropTypes.string,
+    okapi: PropTypes.object,
   };
 
+  static defaultProps = {
+    totalPaidAmount: '',
+    owedAmount: '',
+  }
+
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      actionAllowed: false,
+      accountRemainingAmount: '0.00',
+      prevValidatedAmount: null,
+      prevValidationError: '',
+    };
+
+    this._isMounted = false;
+  }
+
+  componentDidMount() {
+    this._isMounted = true;
+  }
+
+  componentWillUnmount() {
+    this._isMounted = false;
+  }
+
   onClose = () => {
-    const { onClose, reset } = this.props;
+    const { onClose, form: { reset } } = this.props;
+    this._isMounted = false;
 
     onClose();
     reset();
@@ -86,11 +85,13 @@ class ActionModal extends React.Component {
     const {
       accounts = [],
       action,
-      currentValues: { amount },
+      form: { getState },
       intl: { formatMessage },
     } = this.props;
 
-    const selected = calculateSelectedAmount(accounts);
+    const { values: { amount } } = getState();
+
+    const selected = calculateSelectedAmount(accounts, this.isRefundAction(action));
     const type = parseFloat(amount) < parseFloat(selected)
       ? formatMessage({ id: `ui-users.accounts.${action}.summary.partially` })
       : formatMessage({ id: `ui-users.accounts.${action}.summary.fully` });
@@ -136,7 +137,10 @@ class ActionModal extends React.Component {
   }
 
   renderMethod = (options) => {
-    const { action } = this.props;
+    const {
+      action,
+      intl: { formatMessage }
+    } = this.props;
     return (
       <Col xs={this.isPaymentAction(action) ? 3 : 7}>
         <Row>
@@ -147,16 +151,13 @@ class ActionModal extends React.Component {
         </Row>
         <Row>
           <Col xs id="action-selection">
-            <FormattedMessage id={`ui-users.accounts.${action}.method.placeholder`}>
-              {placeholder => (
-                <Field
-                  name="method"
-                  component={Select}
-                  dataOptions={options}
-                  placeholder={placeholder}
-                />
-              )}
-            </FormattedMessage>
+            <Field
+              name="method"
+              component={Select}
+              dataOptions={options}
+              placeholder={formatMessage({ id: `ui-users.accounts.${action}.method.placeholder` })}
+              validate={this.validateMethod}
+            />
           </Col>
         </Row>
       </Col>
@@ -167,14 +168,94 @@ class ActionModal extends React.Component {
     return action === 'payment';
   }
 
-  onBlurAmount = (e) => {
-    const amount = parseFloat(e.target.value || 0).toFixed(2);
-    e.target.value = amount;
+  isRefundAction = (action) => {
+    return action === 'refund';
   }
 
   onChangeOwner = () => {
-    const { dispatch } = this.props;
-    dispatch(change('payment-many-modal', 'method', null));
+    const { form: { change } } = this.props;
+    change('payment-many-modal', 'method', null);
+  }
+
+  triggerCheckEndpoint = (amount, accountId) => {
+    const {
+      checkAmount,
+      okapi,
+    } = this.props;
+
+    return fetch(`${okapi.url}/accounts/${accountId}/${checkAmount}`,
+      {
+        method: 'POST',
+        headers: {
+          'X-Okapi-Tenant': okapi.tenant,
+          'X-Okapi-Token': okapi.token,
+          'Content-Type': 'application/json',
+        },
+        body: JSON.stringify({ amount })
+      });
+  };
+
+  validateAmount = async (value) => {
+    let error;
+
+    const {
+      actionAllowed,
+      accountRemainingAmount,
+      prevValidatedAmount,
+      prevValidationError,
+    } = this.state;
+
+    if (_.isEmpty(value)) {
+      error = <FormattedMessage id="ui-users.accounts.error.field" />;
+    } else if (value !== prevValidatedAmount && this._isMounted) {
+      const { id } = _.head(this.props.accounts) || {};
+      const response = await this.triggerCheckEndpoint(value, id);
+      const {
+        allowed,
+        errorMessage,
+        remainingAmount,
+      } = await response.json();
+
+      this.setState({ prevValidatedAmount: value });
+
+      if (!_.isUndefined(errorMessage)) {
+        this.setState({ prevValidationError: errorMessage });
+        error = errorMessage;
+      }
+
+      if (actionAllowed !== allowed || accountRemainingAmount !== remainingAmount) {
+        this.setState({
+          actionAllowed: allowed,
+          accountRemainingAmount: remainingAmount,
+        });
+      }
+    } else {
+      error = prevValidationError;
+    }
+
+    return error;
+  }
+
+  validateMethod = (value) => {
+    let error;
+    const { action } = this.props;
+
+    if (!value) {
+      error = <FormattedMessage id={`ui-users.accounts.${action}.error.select`} />;
+    }
+
+    return error;
+  }
+
+  validateComment = (value) => {
+    let error;
+    const { commentRequired } = this.props;
+
+    if (commentRequired && !value) {
+      error = <FormattedMessage id="ui-users.accounts.error.comment" />;
+    }
+
+    return error;
   }
 
   render() {
@@ -182,22 +263,30 @@ class ActionModal extends React.Component {
       accounts,
       action,
       balance,
+      totalPaidAmount,
+      owedAmount,
       commentRequired,
-      currentValues: {
-        amount,
-        notify,
-        ownerId
-      },
+      form: { getState },
+      intl: { formatMessage },
       data,
       feefines,
       handleSubmit,
-      invalid,
       label,
       open,
       owners,
       pristine,
       submitting,
     } = this.props;
+
+    const { accountRemainingAmount } = this.state;
+
+    const {
+      valid,
+      values: {
+        notify,
+        ownerId,
+      }
+    } = getState();
 
     let showNotify = false;
     accounts.forEach(a => {
@@ -208,8 +297,7 @@ class ActionModal extends React.Component {
       }
     });
 
-    const selected = calculateSelectedAmount(accounts);
-    const remaining = amount > 0 ? parseFloat(balance - amount).toFixed(2) : parseFloat(balance).toFixed(2);
+    const selected = calculateSelectedAmount(accounts, this.isRefundAction(action));
     const ownerOptions = owners.filter(o => o.owner !== 'Shared').map(o => ({ value: o.id, label: o.owner }));
 
     let options = (this.isPaymentAction(action)) ? data.filter(d => (d.ownerId === (accounts.length > 1 ? ownerId : (accounts[0] || {}).ownerId))) : data;
@@ -217,6 +305,7 @@ class ActionModal extends React.Component {
 
     return (
       <Modal
+        data-test-fee-fine-action-modal
         id={`${action}-modal`}
         open={open}
         label={<FormattedMessage id={`ui-users.accounts.${action}.modalLabel`} />}
@@ -224,22 +313,34 @@ class ActionModal extends React.Component {
         size="medium"
         dismissible
       >
-        <form>
+        <form onSubmit={handleSubmit}>
           <Row>
             <Col xs>{this.renderModalLabel()}</Col>
           </Row>
           <br />
           <Row>
             <Col xs={5}>
-              <Row end="xs">
-                <Col xs={7}>
-                  <FormattedMessage id="ui-users.accounts.totalOwed" />
-                  :
-                </Col>
-                <Col xs={4}>
-                  {balance}
-                </Col>
-              </Row>
+              { this.isRefundAction(action) ? (
+                <Row end="xs">
+                  <Col xs={7}>
+                    <FormattedMessage id="ui-users.accounts.totalPaid" />
+                    :
+                  </Col>
+                  <Col xs={4}>
+                    {totalPaidAmount}
+                  </Col>
+                </Row>
+              ) : (
+                <Row end="xs">
+                  <Col xs={7}>
+                    <FormattedMessage id="ui-users.accounts.totalOwed" />
+                    :
+                  </Col>
+                  <Col xs={4}>
+                    {balance}
+                  </Col>
+                </Row>
+              ) }
               <Row end="xs">
                 <Col xs={7}>
                   <FormattedMessage id="ui-users.accounts.selectedAmount" />
@@ -263,11 +364,11 @@ class ActionModal extends React.Component {
                       name="amount"
                       component={TextField}
                       hasClearIcon={false}
-                      onBlur={this.onBlurAmount}
                       fullWidth
                       marginBottom0
                       autoFocus
                       required
+                      validate={this.validateAmount}
                     />
                   </div>
                 </Col>
@@ -278,9 +379,20 @@ class ActionModal extends React.Component {
                   :
                 </Col>
                 <Col xs={4}>
-                  {remaining}
+                  { accountRemainingAmount || <NoValue /> }
                 </Col>
               </Row>
+              { this.isRefundAction(action) && (
+                <Row end="xs">
+                  <Col xs={7}>
+                    <FormattedMessage id="ui-users.accounts.otherOwed" />
+                    :
+                  </Col>
+                  <Col xs={4}>
+                    {owedAmount}
+                  </Col>
+                </Row>
+              )}
             </Col>
             {(this.isPaymentAction(action) && accounts.length > 1) &&
               <Col xs={4}>
@@ -292,18 +404,14 @@ class ActionModal extends React.Component {
                 </Row>
                 <Row>
                   <Col xs>
-                    <FormattedMessage id="ui-users.accounts.payment.owner.placeholder">
-                      {placeholder => (
-                        <Field
-                          id="ownerId"
-                          name="ownerId"
-                          component={Select}
-                          dataOptions={ownerOptions}
-                          placeholder={placeholder}
-                          onChange={this.onChangeOwner}
-                        />
-                      )}
-                    </FormattedMessage>
+                    <Field
+                      id="ownerId"
+                      name="ownerId"
+                      component={Select}
+                      dataOptions={ownerOptions}
+                      placeholder={formatMessage({ id: 'ui-users.accounts.payment.owner.placeholder' })}
+                      onChange={this.onChangeOwner}
+                    />
                   </Col>
                 </Row>
               </Col>
@@ -330,6 +438,7 @@ class ActionModal extends React.Component {
                 id="comments"
                 name="comment"
                 component={TextArea}
+                validate={this.validateComment}
               />
             </Col>
           </Row>
@@ -341,7 +450,7 @@ class ActionModal extends React.Component {
                     id="notify"
                     name="notify"
                     component={Checkbox}
-                    checked={notify}
+                    type="checkbox"
                     inline
                   />
                   {' '}
@@ -381,8 +490,8 @@ class ActionModal extends React.Component {
               <Button
                 id="submit-button"
                 buttonStyle="primary"
-                onClick={handleSubmit}
-                disabled={pristine || submitting || invalid}
+                type="submit"
+                disabled={pristine || submitting || !valid || !this.state.actionAllowed}
               >
                 <FormattedMessage id={`ui-users.accounts.${action}`} />
               </Button>
@@ -394,19 +503,9 @@ class ActionModal extends React.Component {
   }
 }
 
-const ActionModalRedux = reduxForm({
-  enableReinitialize: true,
-  validate,
+export default stripesFinalForm({
+  initialValuesEqual: (a, b) => _.isEqual(a, b),
+  navigationCheck: true,
+  subscription: { values: true },
+  mutators: { setFieldData },
 })(ActionModal);
-
-const selector = (form, ...other) => (formValueSelector(form))(...other);
-
-export default connect((state, { form }) => ({
-  currentValues: selector(
-    form,
-    state,
-    'amount',
-    'notify',
-    'ownerId'
-  )
-}))(ActionModalRedux);