+++ src/components/Accounts/Actions/index.js
@@ -1,2 +1,2 @@
 export { default as CancellationModal } from './CancellationModal';
-export { default as Actions } from './Actions';
+export { default as Actions } from './FeeFineActions';