+++ test/bigtest/network/factories/feefine.js
@@ -3,7 +3,8 @@ import faker from 'faker';
 
 export default Factory.extend({
   id: () => faker.random.uuid(),
-  // ownerId: () => faker.random.uuid(),
+  automatic: false,
+  ownerId: () => faker.random.uuid(),
   feeFineType : (i) => 'Damage camera fee' + i,
   defaultAmount : (i) => 1000.0 + i * 10,